Moved the `swift build` command from `RUN` to `CMD` and removed the `ADD` in the `Dockerfile`.
What these did was copy the sources and bake them as well as the build products (of the swift build) into the image when the image at image build time. That's bad, because it means any change in the sources would create a new image layer.

The better way is to map the local sources into the container at runtime (`docker run`) and run the `swift build` command as the container command, i.e. in `CMD`.

Hope that makes sense, let me know if you have questions.